requirements.txt
# Web Framework and GUI
Flask==2.0.1
PyQt5==5.15.4
PyQtWebEngine==5.15.4

# AI/ML Dependencies
transformers==4.15.0
torch==1.9.0
numpy==1.21.2

# Document Processing
Pillow==8.3.2
pytesseract==0.3.8
PyMuPDF==1.19.1
python-docx==0.8.11

# Build Tools
pyinstaller==4.5.1

# Additional Dependencies
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pyparsing==2.4.7
regex==2021.8.28
sacremoses==0.0.45
tokenizers==0.10.3
tqdm==4.62.2